// src/modules/master/users/interface.ts
export interface IUser {
  _id?: string
  role_id?: string
  name?: string
  username?: string
  trimmed_username?: string
  email?: string
  trimmed_email?: string
  new_email?: string | null
  trimmed_new_email?: string | null
  avatar_url?: string
  notes?: string | null
  password?: string
  email_verification?: {
    code?: string | null
    url?: string | null
    requested_at?: Date | null
    is_verified?: boolean
    verified_at?: Date | null
  }
  new_email_verification?: {
    code?: string | null
    url?: string | null
    requested_at?: Date | null
    is_verified?: boolean
    verified_at?: Date | null
  } | null
  request_password?: {
    requested_at?: Date | null
    code?: string | null
    url?: string | null
  } | null
  is_archived?: boolean | null
  created_by_id?: string
  created_at?: Date
  updated_at?: Date | null
  updated_by_id?: string | null
  archived_at?: Date | null
  archived_by_id?: string | null
}

export interface IAuthUser {
  _id?: string
  name?: string
  username?: string
  email?: string
  role?: {
    _id?: string
    code?: string
    name?: string
    permissions?: string[]
  }
}